more oriented object thanks to Request and Response class
